Populated Model
Filled in all the non-low-priority model attributes. The functionality
is basic and will likely be fleshed out to include proper 2-way
association functionality, proper composition behavior, etc. in a
controller class. This will allow the interactions between model members
to be abstracted away from the database.
Basic Model testing must be done before attempting to merge this with
the Dev branch.