Model updates
Switched lists to ICollections to work better with Entity Framework.
Switched one String to a FinancialOfficer.
Added in string trimming to all string set methods.
Decided testing should only be done for the validation side of the
model, and that is not part of this task. So, more testing will occur
when further validation is added.