Added two way association functionality
When setting the account associated with a cheque req, the cheque req is
added to the account's list of cheque reqs.
This automatic addition would be annoying to implement from the other
end (adding a cheque req to an account sets the account as the cheque
req's account), and so this was left out.
Cheque Reqs should always be added to an account by setting the account
field in the cheque req.