Seeding
Changed the AccountType field in LedgerAccount to just the Group Enum
(since it isn't a collection, it can be primitive... and also I was
having issues with the foreign key always being null for some reason.
This fixes it).
Seeded 20 accounts. 19 are real, and one is fake. I used a SERVICES
label for the 3 that Kareem gave me that didn't have a specified
"group". This is subject to chenge in the future.
Gave our test financial officer a list of Account Groups he can access
(the 19 real ones seeded).
Fixed the Foreign Key properties for Account Type. And now that Account
Type is only used for UserProfiles, the foreign key is not Nullable.
Fixed a unit test that used the AccountType for a ledger account. It now
uses just the group enum.
Fixed the method that returns all accessible accounts for a given user
so that it uses the Enum in LedgerAccount instead of AccountType.
Fiddled with the seed file to make sure some Foreign keys are set
properly.