Update group functionality to be editable at runtime
Changed the groups from an enum to a string. The database now has a list
of strings called ValidTypes. This can be used to store all of the valid
groups, and can be cross referenced to check if a new account has a
valid type, or something along those lines.